Move sprite texture stuff into sprite_renderer
